+++ app/Http/Controllers/ProfesorController.php
@@ -8,21 +8,24 @@ use Illuminate\Support\Facades\Gate;
 
 class ProfesorController extends Controller
 {
-    public function index(){
+    public function index()
+    {
         $listaProfesor = Profesor::all();
         $message = "Profesores registrados";
-        
+
         return view('Profesor.list', compact('listaProfesor', 'message'));
     }
 
-    public function create(){
-        if(Gate::allows('create-profesor')){
+    public function create()
+    {
+        if (Gate::allows('create-profesor')) {
             return view('Profesor.create');
         }
-        $message = "Acción restringida";
-       return view('Profesor.notification', compact('message'));
+        $message = "Acción restringida - Ingresar Profesor: Solo usuarios con el rol 1";
+        return view('Profesor.notification', compact('message'));
     }
-    public function store(Request $request){
+    public function store(Request $request)
+    {
         $id = $request->input("p_id");
         $nombre = $request->input('p_nombre');
         $apellido = $request->input('p_apellido');
@@ -33,50 +36,61 @@ class ProfesorController extends Controller
         $profesor->apellido = $apellido;
         $profesor->titulo = $title;
         $affected = $profesor->save();
-        if($affected > 0 ){
+        if ($affected > 0) {
             $message = "Profesor registrado correctamente";
-        }else{
+        } else {
             $message = "Ha ocurrido un error";
         }
         return view('Profesor.notification', compact('message'));
     }
 
-    public function show($id){
+    public function show($id)
+    {
         $profesor = Profesor::find($id);
         return view('Profesor.edit', compact('profesor'));
     }
-    public function destroy($id){
-        $affected = Profesor::find($id)->delete();
-        if($affected > 0 ){
-            $message = "Registro eliminado correctamente";
-        }else{
-            $message = "Ha ocurrido un error";
+    public function destroy($id)
+    {
+        if (Gate::allows('create-profesor')) {
+            $affected = Profesor::find($id)->delete();
+            if ($affected > 0) {
+                $message = "Registro eliminado correctamente";
+            } else {
+                $message = "Ha ocurrido un error";
+            }
+            return view('Profesor.notification', compact('message'));
         }
+        $message = "Acción restringida - Eliminar Profesor: Solo usuarios con el rol 1";
         return view('Profesor.notification', compact('message'));
     }
 
 
-    public function update(Request $request, $id){
-        $nombre = $request->input('p_nombre');
-        $apellido = $request->input('p_apellido');
-        $titulo = $request->input('p_titulo');
-        $profesor = Profesor::find($id);
-        $profesor->nombre = $nombre;
-        $profesor->apellido = $apellido;
-        $profesor->titulo = $titulo;
-        $affected = $profesor->save(); 
-        if($affected > 0){
-            $message = "Registro actualizado correctamente";
-        }elseif($affected == 0){
-            $message = "No han ocurrido cambios";
+    public function update(Request $request, $id)
+    {
+        if (Gate::allows('create-profesor')) {
+            $nombre = $request->input('p_nombre');
+            $apellido = $request->input('p_apellido');
+            $titulo = $request->input('p_titulo');
+            $profesor = Profesor::find($id);
+            $profesor->nombre = $nombre;
+            $profesor->apellido = $apellido;
+            $profesor->titulo = $titulo;
+            $affected = $profesor->save();
+            if ($affected > 0) {
+                $message = "Registro actualizado correctamente";
+            } elseif ($affected == 0) {
+                $message = "No han ocurrido cambios";
+            } else {
+                $message = "Ha ocurrido un error al actulizar el registro";
+            }
+            return view('Profesor.notification', compact('message'));
         }
-        else{
-            $message = "Ha ocurrido un error al actulizar el registro";
-        }
+        $message = "Acción restringida - Actualizar Profesor: Solo usuarios con el rol 1";
         return view('Profesor.notification', compact('message'));
     }
 
-    public function viewRelations($id){
+    public function viewRelations($id)
+    {
         $relation = Profesor::with(['clases', 'aulas'])->where('id', $id)->get()[0];
         $message = "";
         //print_r($relation->nombre);
